# api/API.yaml
openapi: 3.0.3
info:
  version: 3.0.2
  title: Mop API
  description: "A list of the currently provided Interfaces to interact with the cluster, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local mop node main API
      port:
        default: "1633"
        description: Service port provided in mop node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local mop node main API
      port:
        default: "1633"
        description: Service port provided in mop node config
paths:
...
  "/mop":
    post:
      summary: "Upload file or a collection of files"
      description: "In order to upload a collection, user can send a multipart request with all the files populated in the form data with appropriate headers.\n\n
        User can also upload a tar file along with the cluster-collection header. This will upload the tar file after extracting the entire directory structure.\n\n
        If the cluster-collection header is absent, all requests (including tar files) are considered as single file uploads.\n\n
        A multipart request is treated as a collection regardless of whether the cluster-collection header is present. This means in order to serve single files
        uploaded as a multipart request, the cluster-index-document header should be used with the name of the file."
      tags:
        - MOP
      parameters:
        - in: query
          name: name
          schema:
            $ref: "Common.yaml#/components/schemas/FileName"
          required: false
          description: Filename when uploading single file
        - $ref: "Common.yaml#/components/parameters/ClusterTagParameter"
        - $ref: "Common.yaml#/components/parameters/ClusterPinParameter"
        - $ref: "Common.yaml#/components/parameters/ClusterEncryptParameter"
        - $ref: "Common.yaml#/components/parameters/ContentTypePreserved"
        - $ref: "Common.yaml#/components/parameters/ClusterCollection"
        - $ref: "Common.yaml#/components/parameters/ClusterIndexDocumentParameter"
        - $ref: "Common.yaml#/components/parameters/ClusterErrorDocumentParameter"
        - $ref: "Common.yaml#/components/parameters/ClusterVoucherBatchId"
        - $ref: "Common.yaml#/components/parameters/ClusterDeferredUpload"
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
          application/x-tar:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Ok
          headers:
            "cluster-tag":
              $ref: "Common.yaml#/components/headers/ClusterTag"
            "etag":
              $ref: "Common.yaml#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "Common.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "Common.yaml#/components/responses/400"
        "402":
          $ref: "Common.yaml#/components/responses/402"
        "403":
          $ref: "Common.yaml#/components/responses/GatewayForbidden"
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
  "/mop/{reference}":
    get:
      summary: "Get file or index document from a collection of files"
      tags:
        - MOP
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "Common.yaml#/components/schemas/ClusterReference"
          required: true
          description: Cluster address of content
      responses:
        "200":
          description: Ok
          # "cluster-feed-index":
          #   $ref: "Common.yaml#/components/headers/ClusterFeedIndex"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "Common.yaml#/components/responses/400"
        "404":
          $ref: "Common.yaml#/components/responses/404"
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
  "/mop/{reference}/{path}":
    get:
      summary: "Get referenced file from a collection of files"
      tags:
        - MOP
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "Common.yaml#/components/schemas/ClusterReference"
          required: true
          description: Cluster address of content
        - in: path
          name: path
          schema:
            type: string
          required: true
          description: Path to the file in the collection.
      responses:
        "200":
          description: Ok
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

        "400":
          $ref: "Common.yaml#/components/responses/400"
        "404":
          $ref: "Common.yaml#/components/responses/404"
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SecurityToken